fix : Fix SignWidget UI/UX
